:art:replaces old websocket pkg
