--- websocket/incoming.go
@@ -1,196 +1,0 @@
-package websocket
-
-import (
-	"strconv"
-	"strings"
-)
-
-// payloadData ...
-// Deprecated: just use a byte slice or json.RawMessage instead
-type payloadData []byte
-
-// UnmarshalJSON ...
-func (pd *payloadData) UnmarshalJSON(data []byte) error {
-	*pd = payloadData(data)
-	return nil
-}
-
-// ByteArr get the byte slice
-func (pd *payloadData) ByteArr() []byte {
-	return []byte(*pd)
-}
-
-type gatewayEventJSON struct {
-	Op             uint        `json:"op"`
-	Data           payloadData `json:"d"`
-	SequenceNumber uint        `json:"s"`
-	EventName      string      `json:"t"`
-}
-
-func (ge *gatewayEventJSON) gatewayEvent() *gatewayEvent {
-	return &gatewayEvent{
-		Op:             ge.Op,
-		Data:           ge.Data,
-		SequenceNumber: ge.SequenceNumber,
-		EventName:      ge.EventName,
-	}
-}
-
-type gatewayEvent struct {
-	Op             uint        `json:"op"`
-	Data           payloadData `json:"d"`
-	SequenceNumber uint        `json:"s"`
-	EventName      string      `json:"t"`
-}
-
-// GetOperationCode ...
-func (ge *gatewayEvent) GetOperationCode() uint {
-	return ge.Op
-}
-
-// UnmarshalJSON see interface json.Unmarshaler
-func (ge *gatewayEvent) UnmarshalJSON(data []byte) (err error) {
-	var i int
-
-	// t
-	t := []byte{
-		'{', '"', 't', '"', ':',
-	}
-	for i = range t {
-		if t[i] != data[i] {
-			evt := &gatewayEventJSON{}
-			err = unmarshal(data, evt)
-			*ge = *evt.gatewayEvent()
-			return
-		}
-	}
-	i++                 // jump to next char
-	if data[i] == 'n' { // null
-		i += 4 // skip `null`
-	} else {
-		// extract the t value
-		var val strings.Builder
-		i++ // skip `"`
-		for ; data[i] != '"'; i++ {
-			val.WriteByte(data[i])
-		}
-		i++ // skip `"`
-		ge.EventName = val.String()
-	}
-
-	// s
-	i += 2 // skip `,"`
-	if data[i] != 's' {
-		evt := &gatewayEventJSON{}
-		err = unmarshal(data, evt)
-		*ge = *evt.gatewayEvent()
-		return
-	}
-	i += 3              // skip `s":`
-	if data[i] == 'n' { // null value
-		i += 4 // skip `null`
-	} else {
-		// extract the s value
-		var val strings.Builder
-		for ; data[i] != ','; i++ {
-			val.WriteByte(data[i])
-		}
-		var tmp uint64
-		tmp, err = strconv.ParseUint(val.String(), 10, 64)
-		if err != nil {
-			evt := &gatewayEventJSON{}
-			err = unmarshal(data, evt)
-			*ge = *evt.gatewayEvent()
-			return
-		}
-		ge.SequenceNumber = uint(tmp)
-	}
-
-	// op
-	i += 2              // skip `,"`
-	if data[i] != 'o' { // o as in op
-		evt := gatewayEventJSON{}
-		err = unmarshal(data, &evt)
-		*ge = *evt.gatewayEvent()
-		return
-	}
-	i += 4              // skip `op":`
-	if data[i] == 'n' { // null value
-		i += 4 // skip `null`
-	} else if data[i] == '0' {
-		ge.Op = 0
-		i++ // skip 0
-	} else {
-		// extract the op value
-		var val strings.Builder
-		for ; data[i] != ','; i++ {
-			val.WriteByte(data[i])
-		}
-		var tmp uint64
-		tmp, err = strconv.ParseUint(val.String(), 10, 64)
-		if err != nil {
-			evt := &gatewayEventJSON{}
-			err = unmarshal(data, evt)
-			*ge = *evt.gatewayEvent()
-			return
-		}
-		ge.Op = uint(tmp)
-	}
-
-	// data
-	i += 2 // skip `,"`
-	if data[i] != 'd' {
-		evt := &gatewayEventJSON{}
-		err = unmarshal(data, evt)
-		*ge = *evt.gatewayEvent()
-		return
-	}
-	i += 3 // skip `d":`
-	ge.Data = data[i : len(data)-1]
-
-	//fmt.Println("asdas")
-	return
-}
-
-type getGatewayResponse struct {
-	URL    string `json:"url"`
-	Shards uint   `json:"shards,omitempty"`
-}
-
-type helloPacket struct {
-	HeartbeatInterval uint     `json:"heartbeat_interval"`
-	Trace             []string `json:"_trace"`
-}
-
-type readyPacket struct {
-	SessionID string   `json:"session_id"`
-	Trace     []string `json:"_trace"`
-}
-
-// DiscordWSEvent event interface that is used outside the websocket package
-type DiscordWSEvent interface {
-	Name() string
-	Data() []byte
-	Unmarshal(interface{}) error
-}
-
-// Event holds event data generated by Discord after the socket layer is finished with it
-type Event struct {
-	content *gatewayEvent
-}
-
-// Name identifies the event type by a title/name. eg. MESSAGE_CREATE
-func (evt *Event) Name() string {
-	return evt.content.EventName
-}
-
-// Data gives the content of the event (excluding socket info) which can be parsed directly into it's
-// respective data structure in the form of an byte array.
-func (evt *Event) Data() []byte {
-	return evt.content.Data.ByteArr()
-}
-
-// Unmarshal unmarshal the content of the Discord event into a given data structure
-func (evt *Event) Unmarshal(v interface{}) error {
-	return unmarshal(evt.Data(), v)
-}
