--- ws2/testdata/1.json
@@ -1,1 +1,0 @@
-{"t":"MESSAGE_CREATE","s":11,"op":0,"d":{"type":0,"tts":false,"timestamp":"2018-09-20T15:46:24.811000+00:00","pinned":false,"nonce":"492360893762371584","mentions":[],"mention_roles":[],"mention_everyone":false,"member":{"roles":[],"mute":false,"joined_at":"2018-09-05T09:40:42.081000+00:00","deaf":false},"id":"492360894244716554","embeds":[],"edited_timestamp":null,"content":"hhgjghghkghjgjh","channel_id":"486833041486905347","author":{"username":"Anders","id":"228846961774559232","discriminator":"7237","avatar":"9efec0f3698dd34ced629cd0e24a9ef6"},"attachments":[],"guild_id":"486833041486905345"}}
