--- ws2/testdata/2.json
@@ -1,1 +1,0 @@
-{"t":null,"s":null,"op":11,"d":null}
