--- ws2/testdata/3.json
@@ -1,1 +1,0 @@
-{"t":"TYPING_START","s":10,"op":0,"d":{"user_id":"228846961774559232","timestamp":1537458377,"member":{"user":{"username":"Anders","id":"228846961774559232","discriminator":"7237","avatar":"9efec0f3698dd34ced629cd0e24a9ef6"},"roles":[],"mute":false,"joined_at":"2018-09-05T09:40:42.081000+00:00","deaf":false},"channel_id":"486833041486905347","guild_id":"486833041486905345"}}
