--- ws2/testdata/4.json
@@ -1,1 +1,0 @@
-{"t":"PRESENCE_UPDATE","s":9,"op":0,"d":{"user":{"id":"228846961774559232"},"status":"online","roles":[],"nick":null,"guild_id":"486833611564253184","game":null,"activities":[]}}
