--- ws2/testdata/diff-structure.json
@@ -1,1 +1,0 @@
-{"t":null,"op":11,"s":null,"d":null}
