--- ws2/testdata/small.json
@@ -1,1 +1,0 @@
-{"t":null,"s":null,"op":11,"d":null}
